+++ pkg/apis/mysql/v1alpha1/mysqlcluster_types.go
@@ -243,6 +243,10 @@ type VolumeSpec struct {
 	// EmptyDir. And represents the PVC specification.
 	// +optional
 	PersistentVolumeClaim *core.PersistentVolumeClaimSpec `json:"persistentVolumeClaim,omitempty"`
+
+	// KeepAfterDelete specifies whether the PVC should be kept after the MysqlCluster is deleted.
+	// +optional
+	KeepAfterDelete bool `json:"keepAfterDelete,omitempty"`
 }
 
 // QueryLimits represents the pt-kill parameters, more info can be found